ReaderInterface not necessary
