--- src/Library/Benzina/ReaderInterface.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace App\Library\Benzina;
-
-interface ReaderInterface
-{
-    public function get(string $entityName): StreamInterface;
-}
